requirements.txt
django>=5.1,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.4,<4.8
django-filter>=25.2,<25.3
dj-database-url>=2.2,<2.4
psycopg[binary]>=3.2,<3.3
Pillow>=11.3,<12.0
gunicorn>=22.0,<24.0
